+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] - XXXX.XX.XX
+
+### Added
+
+### Fixed
+
++ Handling of `.so` files in package build
+
+### Changed
+
+### Removed
+
 ## [v0.8.0] - 2024.08.13
 
 ### Added