+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 + Updating environment READMEs for accuracy
++ Updating top-level README to simplify installation process
 
 ### Removed